pyproject.toml
[project]
name = "shubble"
version = "3.0.0"
description = "Real-time shuttle tracking for Rensselaer Polytechnic Institute"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    # Core backend
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "httpx>=0.28.1",
    "asyncpg>=0.30.0",
    "SQLAlchemy>=2.0.41",
    "alembic>=1.14.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.7.0",
    "python-dotenv>=1.1.1",
    "redis",
    # Data processing (used by worker for DataFrame caching)
    "numpy",
    "pandas>=2.0.0,<3.0.0",
]

[dependency-groups]
bubble = [
    "google-genai>=1.0.0",
]
ml = [
    "torch",
    "scikit-learn",
    "matplotlib",
    "seaborn",
    "tqdm>=4.67.1",
    "statsmodels",
    "scipy",
]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "requests",
    "greenlet"
]
